# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nynth Launch Gate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CollectionStatus = typeof CollectionStatus[keyof typeof CollectionStatus];


export const CollectionStatus = {
  locked: 'locked',
  live: 'live',
} as const;

export interface Collection {
  /** The collection slug (document ID in Firestore) */
  id: string;
  name: string;
  /** ISO 8601 datetime string */
  launchDatetime: string;
  status: CollectionStatus;
}

export type CollectionWithCountStatus = typeof CollectionWithCountStatus[keyof typeof CollectionWithCountStatus];


export const CollectionWithCountStatus = {
  locked: 'locked',
  live: 'live',
} as const;

export interface CollectionWithCount {
  id: string;
  name: string;
  launchDatetime: string;
  status: CollectionWithCountStatus;
  signupCount: number;
}

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** ISO 8601 datetime string */
  launchDatetime: string;
}

export interface Signup {
  id: string;
  collectionId: string;
  email: string;
  /** @nullable */
  password?: string | null;
  unlocked: boolean;
  createdAt: string;
}

export interface SendUnlockResult {
  sent: number;
  failed: number;
  errors?: string[];
}

export interface ApiError {
  error: string;
}