# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nynth Launch Gate API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionStatus } from './collectionStatus';

export interface Collection {
  /** The collection slug (document ID in Firestore) */
  id: string;
  name: string;
  /** ISO 8601 datetime string */
  launchDatetime: string;
  status: CollectionStatus;
}