# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nynth Launch Gate API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** ISO 8601 datetime string */
  launchDatetime: string;
}