# lib/api-zod/src/generated/types/collectionWithCount.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nynth Launch Gate API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionWithCountStatus } from './collectionWithCountStatus';

export interface CollectionWithCount {
  id: string;
  name: string;
  launchDatetime: string;
  status: CollectionWithCountStatus;
  signupCount: number;
}