# lib/api-zod/src/generated/types/sendUnlockResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nynth Launch Gate API
 * OpenAPI spec version: 0.1.0
 */

export interface SendUnlockResult {
  sent: number;
  failed: number;
  errors?: string[];
}