# lib/api-zod/src/generated/types/signup.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nynth Launch Gate API
 * OpenAPI spec version: 0.1.0
 */

export interface Signup {
  id: string;
  collectionId: string;
  email: string;
  /** @nullable */
  password?: string | null;
  unlocked: boolean;
  createdAt: string;
}